+++ src/components/file-upload/z-file-upload/index.tsx
@@ -210,7 +210,7 @@ export class ZFileUpload {
   renderUploadButton() {
     return [
       this.renderInput(),
-      <z-button-deprecated
+      <z-button
         onClick={() => this.input.click()}
         onKeyPress={(e) => {
           if (e.code == "Space" || e.code == "Enter") {
@@ -224,7 +224,7 @@ export class ZFileUpload {
         ref={(val) => (this.button = val)}
       >
         Allega
-      </z-button-deprecated>,
+      </z-button>,
     ];
   }