+++ src/components/z-skip-to-content/index.spec.ts
@@ -22,7 +22,7 @@ describe("Suite test ZSkipToContent", () => {
 
     expect(page.root).toEqualHtml(`
      <z-skip-to-content class="dark false" links="[{&quot;ariaLabel&quot;: &quot;vai al link1&quot;, &quot;label&quot;:&quot;vai al link1&quot;, &quot;href&quot;:&quot;#link1&quot;}]" variant="dark">
-       <z-link aria-label="vai al link1" class="link-visible t-weight-sb" href="#link1" id="skip-to-content-0" textcolor="white" underline="">
+       <z-link aria-label="vai al link1" class="link-visible body-1-sb" href="#link1" id="skip-to-content-0" textcolor="white" underline="">
           vai al link1
         </z-link>
       </z-skip-to-content>