+++ src/components/z-table/z-table-header/index.tsx
@@ -152,14 +152,13 @@ export class ZTableHeader {
               position={PopoverPosition["below-center"]}
               background-color={"gray200"}
             >
-              <z-button-deprecated
+              <z-button
                 icon="contextual-menu"
                 variant={ButtonVariantEnum["tertiary"]}
                 size={ButtonSizeEnum["x-small"]}
-                square
                 slot="trigger"
                 onClick={() => this.handleMenuClick()}
-              ></z-button-deprecated>
+              ></z-button>
 
               <div slot="popover">
                 <slot name="contextual-menu" />